do not highlight matching parentheses
:help DoMatchParen

By default the plugin will highlight both the paren under the cursor and the
matching one using the |hl-MatchParen| highlighting group.  This may result in
the cursor briefly disappearing from the screen as the MatchParen colors take
over the cursor highlight.  To prevent this from happening and have the plugin
only highlight the matching paren and not the one under the cursor
(effectively leaving the cursor style unchanged), you can set the
"matchparen_disable_cursor_hl" variable: >

	:let matchparen_disable_cursor_hl = 1

The characters to be matched come from the 'matchpairs' option.  You can
change the value to highlight different matches.  Note that not everything is
possible.  For example, you can't highlight single or double quotes, because
the start and end are equal.